refactor and added screen component
