+++ index.html
@@ -7,7 +7,7 @@
     <title>Simple Egg Timer</title>
   </head>
   <body>
-    <div id="root" class="max-w-[640px] mx-auto"></div>
+    <div id="root"></div>
     <script type="module" src="/src/main.tsx"></script>
   </body>
 </html>
