+++ src/App.tsx
@@ -1,32 +1,18 @@
 import EggPicker from './components/eggPicker'
 import Timer from './components/Timer'
 import useAppStore from './store/useAppStore'
+import Screen from './ui/Screen'
+import Start from './components/Start'
 
 function App() {
   const view = useAppStore((state) => state.view)
-  const setView = useAppStore((state) => state.setView)
 
   return (
-    <main className="relative h-[100dvh] bg-[url('/background.jpg')] bg-cover bg-center">
-      <div className="absolute inset-0 bg-amber-200/90 z-0"></div>
-
-      {view === 'timer' && <Timer />}
-
-      {view === 'menu' && <EggPicker />}
-
-      {view === 'start' && (
-        <div className="absolute inset-0 flex flex-col items-center justify-center gap-5 z-10">
-          <h1 className="font-bold text-2xl">Let's time your egg!</h1>
-          <img className="animate-spin-slow w-124" src="./egg.png" alt="egg" />
-          <button
-            className="px-4 py-2 bg-amber-200 rounded-lg shadow mt-4"
-            onClick={() => setView('menu')}
-          >
-            Start
-          </button>
-        </div>
-      )}
-    </main>
+    <Screen>
+      <>{view === 'timer' && <Timer />}</>
+      <>{view === 'menu' && <EggPicker />}</>
+      <>{view === 'start' && <Start />}</>
+    </Screen>
   )
 }
